new features
created the new method to shoot in the teleop and autonomous, using the SequentialCommandGroup, using this i can use the commands in auto removing the waitCommand in PathPlanner, the auto angle is done but isn't perfect, now you can angle the shooter pressing the X Button, and it angle using the InterpolatingDoubleTreeMap. The PID of the rotation and translation is tuned, but i can do better.